//crud操作

+++ src/main/java/com/xm/demo/domain/Video.java
@@ -3,22 +3,22 @@ package com.xm.demo.domain;
 
 public class Video {
 
-  private long id;
+  private Integer id;
   private String title;
   private String summary;
   private String coverImg;
-  private long viewNum;
-  private long price;
+  private Integer viewNum;
+  private Integer price;
   private java.sql.Timestamp createTime;
-  private long online;
-  private double point;
+  private Integer online;
+  private Double point;
 
 
-  public long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
 
@@ -50,20 +50,20 @@ public class Video {
   }
 
 
-  public long getViewNum() {
+  public Integer getViewNum() {
     return viewNum;
   }
 
-  public void setViewNum(long viewNum) {
+  public void setViewNum(Integer viewNum) {
     this.viewNum = viewNum;
   }
 
 
-  public long getPrice() {
+  public Integer getPrice() {
     return price;
   }
 
-  public void setPrice(long price) {
+  public void setPrice(Integer price) {
     this.price = price;
   }
 
@@ -77,20 +77,20 @@ public class Video {
   }
 
 
-  public long getOnline() {
+  public Integer getOnline() {
     return online;
   }
 
-  public void setOnline(long online) {
+  public void setOnline(Integer online) {
     this.online = online;
   }
 
 
-  public double getPoint() {
+  public Double getPoint() {
     return point;
   }
 
-  public void setPoint(double point) {
+  public void setPoint(Double point) {
     this.point = point;
   }
